requirements.txt
anyio==3.6.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.87.0
h11==0.14.0
idna==3.4
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
Pillow==9.3.0
pydantic==1.10.2
pyparsing==3.0.9
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
torch==1.13.0
torchmetrics==0.10.2
torchvision==0.14.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
